refactor: separate BoardsComponentWrite logic into hook, types, and queries files
