+++ 03-maincamp-triptalk/03-CNA/src/components/boards-write/types.ts
@@ -1,0 +1,10 @@
+export interface BoardsWriteProps {
+  isEdit: boolean;
+  boardId?: string;
+  data?: {
+    fetchBoard?: {
+      title?: string;
+      contents?: string;
+    };
+  };
+}
